Looks like we didn't need to keep an array list of changed entities, which is a good thing.  The boolean will be good enough.  Everything that we notify just repulls whatever entities they need, we never have to send them a copy of the changed entities.  I'm not sure why we thought we might need to...somethign seems fishy.
As a note: The combo box never shows anything.

I added the few lines of renderer updating, but I'm not sure if I paused and resumed the thread correctly, and have no way to test with the combo box not working.

Things almost seem to be getting too simple.  I'm scared....hold me.